+++ Questionable/Controller/Steps/Interactions/EquipRecommended.cs
@@ -88,7 +88,7 @@ internal static class EquipRecommended
                         {
                             chatGui.Print("正在穿上推荐装备（一键最强）", CommandHandler.MessageTag, CommandHandler.TagColor);
                             recommendedEquipModule->EquipRecommendedGear();
-                            _continueAt = DateTime.Now.AddSeconds(1);
+                            _continueAt = DateTime.Now.AddSeconds(0.25);
                         }
 
                         _checkedOrTriggeredEquipmentUpdate = true;
@@ -105,7 +105,7 @@ internal static class EquipRecommended
                         {
                             stylist.UpdateGearset();
                             _checkedOrTriggeredEquipmentUpdate = true;
-                            _continueAt = DateTime.Now.AddSeconds(1);
+                            _continueAt = DateTime.Now.AddSeconds(0.25);
                             return ETaskResult.StillRunning;
                         }
                     }