+++ Questionable/Windows/QuestComponents/ActiveQuestComponent.cs
@@ -677,8 +677,8 @@ internal sealed partial class ActiveQuestComponent
 
     private static string Shorten(string text)
     {
-        if (text.Length > 35)
-            return string.Concat(text.AsSpan(0, 30).Trim(), ((SeIconChar)57434).ToIconString());
+        if (text.Length > 30)
+            return string.Concat(text.AsSpan(0, 25).Trim(), ((SeIconChar)57434).ToIconString());
 
         return text;
     }